// packages/example/src/generated/music-catalog.schemas.gen.ts
export const schemas = {
  "albums": {
    "$schema": "https://json-schema.org/draft/2020-12/schema",
    "$id": "urn:jsonapi:schema:albums",
    "type": "object",
    "description": "An `albums` resource object.",
    "properties": {
      "type": {
        "type": "string",
        "const": "albums"
      },
      "id": {
        "type": "string"
      },
      "attributes": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
// ...
        }
      },
      "relationships": {
        "type": "object"
      },
      "links": {
        "type": "object"
      },
      "meta": {
        "type": "object"
      }
    },
    "required": [
      "type",
      "id"
    ]
  },
// ...
} as const